Much smarter MySQL lock contention handling

+++ site/pages/modules/chat.php
@@ -423,7 +423,8 @@ class Module_chat
 			delete_chatroom($old['id']);
 		}
 		// Prune chat events
-		$GLOBALS['SITE_DB']->query('DELETE FROM '.$GLOBALS['SITE_DB']->get_table_prefix().'chat_events WHERE e_date_and_time<'.strval(time()-CHAT_EVENT_PRUNE));
+		if (!$GLOBALS['SITE_DB']->table_is_locked('chat_events'))
+			$GLOBALS['SITE_DB']->query('DELETE FROM '.$GLOBALS['SITE_DB']->get_table_prefix().'chat_events WHERE e_date_and_time<'.strval(time()-CHAT_EVENT_PRUNE));
 
 		enter_chat_lobby();
